a start - relies on #4481 - untested - still wip
